Disable joining team via console.
If someone connects with a color other than white, the server forces them on a team if the server is in the 'waiting' state. This also allows the player to use the console and change colors to join the team automatically, though it seems more of an annoyance than a convenience.